+++ backend/prisma/migrations/20260917005458_add_map_global_illumination/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "Map" ADD COLUMN     "globalIllumination" BOOLEAN NOT NULL DEFAULT true;
+